code fixes based on PR feedback
